+++ CHANGELOG.md
@@ -10,8 +10,12 @@
 - deps upgrade:
     - filippo.io/age v1.1.1
     - fyne.io/fyne v2.3.0
-    - golang.org/x/crypto v0.5.0
-    - golang.org/x/text v0.6.0
+    - golang.org/x/crypto v0.14.0
+    - golang.org/x/sync v0.1.0
+    - golang.org/x/term v0.13.0
+    - golang.org/x/text v0.13.0
+- deps remove:
+    - github.com/mikesmitty/edkey
 
 ## 0.19.1 - 01 October 2022